Explanatory comments, move build prepeg fn.
