+++ slidechain/import.go
@@ -26,6 +26,9 @@ func (c *Custodian) buildImportTx(
 	buf := new(bytes.Buffer)
 	fmt.Fprintf(buf, "{'C', x'%x', x'%x',", createTokenSeed[:], consumeTokenProg)
 	fmt.Fprintf(buf, " {'Z', %d}, {'T', {x'%x'}},", int64(1), recipPubkey)
+	// For a slight optimization, the anchor for that contract's value is
+	// split from the value generated by the `nonce` instruction. Reconstructing
+	// this new anchor is below.
 	nonceHash := UniqueNonceHash(c.InitBlockHash.Bytes(), expMS)
 	snapshotNonceHash := txvm.VMHash("Split2", nonceHash[:])
 	fmt.Fprintf(buf, " {'V', %d, x'%x', x'%x'},", 0, zeroSeed[:], snapshotNonceHash[:])
